添加转换警告并解决

+++ Middlewares/ST/STM32_USB_Device_Library/Core/Inc/usbd_def.h
@@ -329,7 +329,7 @@ __STATIC_INLINE uint16_t SWAPBYTE(uint8_t *addr)
   _pbuff++;
   _Byte2 = *(uint8_t *)_pbuff;
 
-  _SwapVal = (_Byte2 << 8) | _Byte1;
+  _SwapVal = (uint16_t)((_Byte2 << 8) | _Byte1);
 
   return _SwapVal;
 }
